requirements.txt
flask
flask-cors
rembg
pillow
python-multipart
